requirements.txt
mysql-connector-python
pymysql
paho-mqtt
pyyaml
sqlalchemy